Refactor functions task creation
This commit introduces the minimal amount of changes so that the functions tasks
are created by declaring some of their inputs, which are:

- the runtime classpath of the function
- the archive file

There are many other changes to make, since the tasks do not declare all
inputs, nor declare any outputs. This isn't great for interaction with
other plugins, but this change at least makes it possible for other
plugins to create function tasks which use a different archive file
and/or classpath.