Add support for multiple functions
This commit makes it possible to register multiple functions in a single
project. Basically, instead of assuming that the artifact to be deployed
as a function comes from the `jar` task, it makes it possible to register
additional functions and configure their classpath accordingly.

Because of backwards compatibility, this does _not_ remove the `azurefunctions`
extension, which configures the default, main function. However, there's now
a container named "azureBinaries" which lets the user (or plugins) register
additional functions. The plugin will take care of creating the corresponding
tasks.